Remove reference to parameter in default value
ISO-14882:2014 (and as far as I can tell also the prior versions)
forbids the use of function parameters in the initializer-clause of a
parameter-declaration (see dcl.fct.default). Even though this general
ban was lifted a little bit - by allowing usage in non-evaluated
contexts - in CWG2082, this change does not apply here.